Update systick and sensor reading
Systick runs at 500Hz and does encoders so far.
At end of systick, sensor read cycle is fired off and all channels read.
Moved ADC related stuff to sensors_control.cpp.
