+++ robot_config_sample.h
@@ -1,6 +1,7 @@
 #ifndef ROBOT_CONFIG_H
 #define ROBOT_CONFIG_H
 
+#include <arduino.h>
 /*******************************************************************
  *
  * Copy this file as robot_config.h in the working directory
@@ -58,8 +59,6 @@ const float DEG_PER_COUNT = (360.0 * MM_PER_COUNT) / (PI * WHEEL_SEPARATION);
 // using voltage based control.
 const float MAX_MOTOR_VOLTS = 6.0f;
 
-const int LOOP_INTERVAL_MS = 10;
-const float LOOP_INTERVAL = (float)LOOP_INTERVAL_MS / 1000.0f;
-const float LOOP_FREQUENCY = 1.0 / LOOP_INTERVAL;
-
+const float LOOP_FREQUENCY = 500.0; //Hz
+const float LOOP_INTERVAL = 1000.0f / LOOP_FREQUENCY;
 #endif
